renamed requestBS4.py -> parcing.py, parcing (2).ipynb -> research.ipynb, for_github. png -> tableau_screen.png
